Fix a bug that has been causing tests/hard_coded/sub-modules/non_word_mutable
Estimated hours taken: 3
Branches: main

Fix a bug that has been causing tests/hard_coded/sub-modules/non_word_mutable
to fail in high-level C grades.

The bug occurred when we had a mutable declaration in the child module whose
type was a foreign type defined in the parent module and where that foreign
type relied on other C type definitions in a foreign_decl pragma in the
parent.  The problem was that the contents of the parent's foreign_decl pragma
were not visible in the child's .mih file at the point where the declarations
for the mutable variables were made.

The fix is to make sure that we #include all ancestor .mih files before any
foreign code declarations made by the child's .mih file.

compiler/mlds_to_c.m:
	Avoid the above problem.

compiler/modules.m:
	Update module qualifier syntax in a comment.