Fix the failure of tests/debugger/save.  The problem was that the on systems
Estimated hours taken: 0.2
Branches: main, release

Fix the failure of tests/debugger/save.  The problem was that the on systems
where the XML browser was enabled the output of this test case included
the commands used to control the XML browser.  These commands will contain
system specific parts.  The fix is to (a) provide a second expected output
that includes the XML browser commands and (b) pipe the output through
sed to remove the system specific bits.

tests/debugger/Mmakefile:
	Pipe the output of the save test case through sed to remove
	system specific dependencies.

tests/debugger/save.exp2:
	Provide a second expected output for systems where the XML
	browser is enabled.