Fix a bug with profiling and the high-level C grades.
Estimated hours taken: 1
Branches: main, release

Fix a bug with profiling and the high-level C grades.  The problem was that we
were only emitting an INIT comment whenever there was a user initialisation or
finalisation predicate.  This is not noticeable in most of the hl* grades
since mercury__<modulename>__init is empty anyway.  However in the profiling
grades the init function is responsible for calling MR_init_entry for each
procedure.  Failing to do this causes unknown functions to show up in the
profile.  (This problem appears to have been introduced with r1.101 of
util/mkinit.c - prior to that, in the 0.12 release for example, init functions
in the hl* grades were handled differently.)

compiler/mlds_to_c.m:
	Always output the INIT comment for a module.