For each module generate a function mercury__<modulename>__required_init
Estimated hours taken: 2
Branches: main, release

For each module generate a function mercury__<modulename>__required_init
that calls any initialisers in the module.  We only create this function if
there are any user-defined initialisation predicates (or any introduced for
mutable initialisation.) This means that at most one REQUIRED_INIT comment per
module will be required, as opposed to the one per initialise or mutable
declaration we have now.

This change goes part of the way to fixing the bug with initialisers
in libraries not being run.  (The remainder of the fix is a separate change.)
This change also makes it simpler to ensure that the initialisers get executed
in the correct order.

Do the same thing for finalisers.

compiler/llds_out.m:
compiler/mlds_to_c.m:
	Make the above changes to the LLDS and MLDS backends respectively.