Fix the bug with initialisers/finalisers in libraries not being called ( and
Estimated hours taken: 7
Branches: main, release

Fix the bug with initialisers/finalisers in libraries not being called ( and
as consequence also fixes the bug with mutables not being given their correct
initial value).  The problem was that the directives necessary to call them
were not being included in the libraries' .init file.

The fix is to add a new mode of operation to mkinit that given a list of
.c files that make up some Mercury library, constructs the .init file for
that library.  In particular, it now constructs the .init file so that
it contains any REQUIRED_{INIT,FINAL} directives needed by the library.
The new mode of operation is invoked when mkinit is given the `-k' option.

Changes to the build systems to use the new mode of operation will be committed
after the changes to mkinit have bootchecked.

util/mkinit.c:
	Add a new mode of operation, `-k'.  When invoked with this
	option mkinit will read a list of .c files and emit a sequence of
	INIT, REQUIRED_INIT and REQUIRED_FINAL directives to stdout,
	corresponding to INIT, REQUIRED_INIT and REQUIRED_FINAL comments
	in the .c files.