Fix the failure of debugger/declarative/find_origin in .debug grades.
Estimated hours taken: 0
Branches: main, release

Fix the failure of debugger/declarative/find_origin in .debug grades.
The problem was that the version of the input and expected output for
those grades had not been updated to take account of the removal of the
set command.

tests/debugger/declarative/find_origin.{inp2,exp2}:
	Update the input and expected output for these test cases.