Fix a bug reported by Greg Duck.  On x86 systems mmc --make was not building
Estimated hours taken: 6
Branches: main, release

Fix a bug reported by Greg Duck.  On x86 systems mmc --make was not building
static executables by default, but was linking in shared versions of some
libraries.  This won't work correctly on x86 machines because of our use of the
register ebx as a global register in non-pic code.  The following patch
causes mmc --make to invoke the linker with the same default settings as mmake.

scripts/Mercury.config.in:
	The default values for the `--linkage' and `--mercury-linkage'
	options on x86 should be `static' not `shared'.