This diff implements step 1 of the fix for a bug I introduced when turning
Estimated hours taken: 2
Branches: main

This diff implements step 1 of the fix for a bug I introduced when turning
the primitives involved in implementing trace goals into builtins. The bug
is that that builtin primitives have to be declared, but if they don't have
definitions, then add_pred *creates* a definition for them which assumes
them to be pure. The primitives involved in trace goal implementations
are semipure and impure respectively.

compiler/add_pred.m:
	Add special case code for adding the definitions of the
	trace_get_io_state and trace_set_io_state primitives. These exploit
	the fact that I/O states are dummy values.

compiler/builtin_ops.m:
	Fix parentheses in a comment.

compiler/options.m:
	Provide a means to test in configure.in whether the installed compiler
	contains this bug fix. This will be used by step 2, which can happen
	only when this diff has been installed on all our machines.