Make accessing the globals field in the I/O state properly thread safe by
Estimated hours taken: 2
Branches: main

Make accessing the globals field in the I/O state properly thread safe by
applying a variation of the transformation recently proposed for mutables on
mercury-developers.

Add a predicate that allows the globals field to be updated atomically in the
presence of multiple threads.

library/io.m:
	Attach a mutex to ML_io_user_globals and use that govern access to the
	globals field in the I/O state.

	Add a predicate the allows the globals field to be updated atomically.

NEWS:
	Mention the new addition.