This diff implements step 2 of the fix for a bug I introduced when turning
Estimated hours taken: 0.5
Branches: main

This diff implements step 2 of the fix for a bug I introduced when turning
the primitives involved in implementing trace goals into builtins. The bug
is that that builtin primitives have to be declared, but if they don't have
definitions, then add_pred *creates* a definition for them which assumes
them to be pure. The primitives involved in trace goal implementations
are semipure and impure respectively.

library/private_builtin.m:
	Add declarations for trace_get_io_state and trace_set_io_state.

configure.in:
	Require the installed compiler to have part 1 of this fix.