Do not use the installed mmc to compile the slice directory, because
Estimated hours taken: 0.5
Branches: main

Do not use the installed mmc to compile the slice directory, because
this means you cannot install the source distribution on a machine that
doesn't already have Mercury installed.

The reason the slice directory was compiled with the installed mmc was so that
the mtc_union tool wouldn't be compiled with debugging enabled even if the
compiler was compiled with debugging.  mtc_union is used to summarise
trace counts when coverage testing the compiler.  If mtc_union is compiled
in a debug grade then tail recursion is lost and it runs out of stack space.

If we're going to coverage test the compiler then we should use the
MERCURY_OPTIONS option that lets us customise the summary command, to choose
a version of mtc_union not compiled with debugging.

Mmakefile:
	Remove PLAIN_SUBDIR_MMAKE and instead use SUBDIR_MMAKE to compile
	the slice directory.

configure.in:
	Build SLICE_FLAGS.

slice/Mercury.options:
	Don't warn about too lax determinism declarations in mtc_diff
	(some of the option handling predicates have a multi determinism
	declaration when they're actually det, since the option handling
	predicates require a multi predicate).

slice/Mmakefile:
	Revert 1.5 changes and add rules for mtc_diff and mcov.

slice/SLICE_FLAGS.in:
	Put back flags for slice directory.