Fix a bug with intermodule optimization and termination analysis.
Estimated hours taken: 3
Branches: main, release

Fix a bug with intermodule optimization and termination analysis.  The problem
was that not all of the termination_info pragmas for all of the opt_exported
procedures were being written to the .opt file.  In some cases where the
procedure was defined by a foreign clause, this lead to the termination
foreign code attributes conflicting with the analyser's default assumptions
about procedures without termination_info pragmas.

This diff fixes the bug in one direction: making sure that we always adjust
the import status of items written to .opt files.  This is sufficient to get
the termination analyser to write out the missing termination_info pragmas.
(Previously we used to only do this if intermodule unused argument analysis
was enabled for some reason.)

I'm working on a separate fix in the other direction, i.e. making the
termination analyser a bit more robust with respect to missing information -
I'll post that as a separate diff.  (In any case this change is useful in it's
own right - it should improve accuracy).

compiler/intermod.m:
 	Always adjust the import status of items written to .opt files.