Fix a bug with parametric modes.  Mode specific clauses or foreign_export
Estimated hours taken: 1
Branches: main, release

(For the release branch s/foreign_export/export/)

Fix a bug with parametric modes.  Mode specific clauses or foreign_export
pragmas that involved a parametric mode didn't work because the compiler
could not match the clause or foreign_export pragma with the corresponding
mode declaration.  The problem here is that the inst varsets attached to
the mode declaration and clause or foreign_export pragma are not necessarily
the same so unifying the modes is not sufficient to prove that they match.

Since we have already fixed this problem for foreign_proc pragmas the fix here
is just to extend that solution to mode-specific clauses and foreign_export
pragmas.  The fix is just to allow for a renaming between the inst variables.
(Note: I've also extended this fix to the termination_info and structure
sharing/reuse pragmas which could also be affected by this.)

compiler/add_pragma.m:
 	Use get_procedure_matching_declmodes_with_renaming/4 instead
 	of get_procedure_matching_declmodes/4.  This fixes problems
 	with items that contain inst variables not matching the
 	corresponding mode declaration for that item.

 	Delete get_procedure_matching_declmodes/4 as it isn't used
 	anymore and is fundamentally broken w.r.t to the way inst
  	variables are currently handled inside the compiler.

 	Move some code around so that the predicates are in top-down
 	order.

compiler/add_clause.m:
 	Call get_procedure_matching_declmodes_with_renaming/4 instead
 	of get_procedure_matching_declmodes.

tests/valid/Mmakefile:
tests/valid/param_mode_bug.m:
 	Add a test case for the above.