Fix a bug with mmc --make and static libraries.
Estimated hours taken: 1
Branches: main, release

Fix a bug with mmc --make and static libraries.  We need to update the
archive index after copying the library to the installation directory.
The linker on Mac OS X complains if we don't do this.  This is also the
behaviour of mmake anyway, so mmc --make should do the same.

compiler/make.program_target.m:
	Rebuild the archive index after installing a static library.