Add a new module to the standard library that provides generic stream
Estimated hours taken: 30 (+ 15 by Ian) (+ unknown late last year)
Branches: main

Add a new module to the standard library that provides generic stream
handling via a family of typeclasses.  This subsumes the functionality
provided by the existing stream library in extras and improves upon that
design in several ways.

This incarnation of streams is parameterized over the state that is
updated by the stream operations, i.e. it is no longer restricted to
just being the I/O state.  Streams are also parameterized over the
type of the data being written to, or read from, them.  This allows for
efficient implementations of operations like write_string, and also
allows the put and get methods for a single stream to be overloaded.

Subsequent diffs will extend the io module to make the standard file
streams instances of these typeclasses and convert things like the
pretty printer and term_to_xml to use them as well.

library/stream.m:
	New file.  This module defines the stream typeclasses.

library/library.m:
	Include the new library module.