Fix some problems with my recent change to simplify to optimize the handling
Estimated hours taken: 6
Branches: main

Fix some problems with my recent change to simplify to optimize the handling
of calls to the functions in int.m involving the bits_per_int platform-specific
constant.

One problem was that the transformation in simplify.m worked only if it could
find in the HLDS's predicate table the functions it needed (unchecked_quotient,
unchecked_remainder and *), but these were deleted earlier as dead predicates.
The fix is to prevent dead_proc_elim from deleting these predicates. (Three
small predicates aren't a problem for memory consumption.)

Another problem was that some of the calls to the library predicates that
simplify was trying to optimize were introduced as part of intermodule inlining
*after* the first simplify pass, and inlining also replaced the calls with
their definitions, which are foreign_procs. The fix is to apply the same
transformation to the foreign_proc form of the call as to the plain_call form.

compiler/dead_proc_elim.m:
	Do not eliminate predicates that simplify may introduce calls to.

	Rename several predicates to avoid ambiguities, and some function
	symbols and variables to make their purpose clearer.

	Instead of printing out warning messages on the fly, return them to be
	printed all at once by mercury_compile.m.

compiler/simplify.m:
	Export a predicate that dead_proc_elim can use to find out which
	predicates may have calls introduced to them.

	Implement the fix of the second problem noted above.

compiler/mercury_compile.m:
compiler/inlinig.m:
	Conform to the changes in dead_proc_elim.

compiler/error_util.m:
	Add a phase for dead_proc_elim, since that phase can now also generate
	messages.

compiler/pred_table.m:
	Improve some documentation.