This fix is due to Jon Morgan.
Estimated hours taken: 2.5
Branches: main

Fix a bug in the IL backend, which appeared in the compilation of
injection.delete_values with inlining enabled.  The code generator was
sometimes generating functions and predicates that only differed in
the return type when inlining higher order terms.  This caused
duplicate method errors in the generated IL.  The fix is to alter the
name mangling routine so that it always distinguishes between
functions and predicates.

compiler/mlds_to_il.m:
	Add "_f" to all functions when mangling MLDS procedure names.