Fix a bug that was causing `make install' to break when building
Estimated hours taken: 3.5
Branches: main

Fix a bug that was causing `make install' to break when building
Mercury from the source distribution.  The problem was that the set of
.int0 files that the target `mer_std.install_ints' depended on contained
.int0 files that were not actually built during a normal build.  At
installation time mmake tried building these "missing" files but if, as
was likely when installing from the source distribution, a Mercury compiler
was not available building them would fail.

The reason that the set of .int0 files was inaccurate was actually a
workaround for another bug - namely that the compiler generates unnecessary
.int0 files for nested sub-modules with no children.  The workaround was
intended to ensure that realclean target deleted any of these spurious .int0
files.

This diff restores the code that writes the accurate set of .int0 files
to the .dv file.  As a workaround for the existing problem with the nested
sub-modules we modify the realclean target to assume that every module in the
program potentially has a private interface.

compiler/modules.m:
	Make the changes described above.

	Unrelated changes: switch on the target_lang and foreign_lang
	types rather than using them in if-then-elses.  This makes it
	easier to spot where to add new target and foreign languages.

	Handle one such new target, x86_64, correctly in a few spots.