Fix a bug that was causing hard_coded/copy_pred_2 to fail on 64-bit machines
Estimated hours taken: 6
Branches: main

Fix a bug that was causing hard_coded/copy_pred_2 to fail on 64-bit machines
in hlc* grades.  The problem was that the MLDS->C code generator was
inconsistent in how it mapped the mlds_native_{int,float,char}_type types to
C, e.g. sometime mlds_native_int_type was an MR_Integer and sometimes it was
an int.

On machines where MR_Integer is not an int, e.g. saturn where it is a long,
this caused the test case to fail.  The actual test case failure was caused
by the fact that typevar_vector array in the closure layout was declared to
be an array of ints while the compiler was trying to put values of type
MR_Integer into it.  This didn't work since (on saturn) the array was half
the size it needed to be.  (The actual test case failure was an assertion
failure that occurred when deep copy was trying to use such a closure layout
to make a copy of a closure.)

compiler/mlds_to_c.m:
	Fix an inconsistency with how we map MLDS types to C types.