Optimize code sequences such as
Estimated hours taken: 1
Branches: main

Optimize code sequences such as

	mkframe(framesize, do_fail)
	...
	if (...) goto ...
	...
	succeed()

in which the ...s do not contains labels or update any of the control slots
of the frame. The optimization is replacing the succeed with a succeed_discard,
avoiding a future backtrack that cannot succeed.

compiler/peephole.m:
	Look for the pattern and act on it.

	Rejig the arguments of peephole's main predicate to avoid the need to
	rebuild the original instruction on the heap by passing it whole,
	instead of in pieces.

	Rename predicates to avoid the need for module qualification.

compiler/opt_util.m:
	Add a utility predicate needed by peephole.

compiler/optimize.m:
	Conform to the predicate renames in peephole.