Fix a bug reported by Ondrej Bojar.  If there were multiple foreign clauses
Estimated hours taken: 1
Branches: main

Fix a bug reported by Ondrej Bojar.  If there were multiple foreign clauses
in the same language for a procedure we silently ignored all but the first;
we now emit an error message.

This isn't a complete fix: multiple foreign clauses in a language not
supported as a foreign language by the current target language won't be
reported, and whether multiple clauses in a foreign language that is
supported by the target language but is not the preferred foreign language
depends upon the presence of clauses in the preferred foreign language (and
their order relative to each other.) Fixing these two problems (particularly
the first) would be a lot more work; given the rarity of the problem it
also has a rather low priority.

compiler/add_pragma.m:
	Emit an error message if there are multiple foreign clauses
	in the same language for a procedure.

tests/invalid/Mmakefile:
tests/invalid/fp_dup_bug.{m,err_exp}:
	Test case for the above.