Improve support for calling procedures in Mercury libraries from
Estimated hours taken: 20
Branches: main

Improve support for calling procedures in Mercury libraries from
applications written in foreign languages, i.e where the program entry point
is not the Mercury predicate main/2.  The main trick is to ensure that the
necessary runtime initialisation is done before any Mercury procedures are
called (or at least to provide the mechanism to do such initialisation,
ensuring that it is done is the programmer's responsibility.)

We currently support this sort of thing via the compiler's `--no-main'
option.  This diff adds a more user friendly mechanism (at least on the
Mercury side.)  In particular, we no longer require that one of the Mercury
libraries define a main/2 predicate.  (Note: the existing behaviour of
--no-main is unchanged.)

For the set of Mercury libraries that we wish to use from a foreign
application we create a standalone interface.  A standalone interface is a
cut-down version of the _init.c file that would be created for a Mercury
executable that uses the same set of libraries.  (Which libraries to include
in the interface can be specified via the usual mechanisms, e.g. the --ml
option.)

The standalone interface has two parts: an object file that contains the
cut-down version of the _init.c file, and a header file that contains the
declarations for the functions that initialise and shut down the Mercury
runtime.  This header file is compatible with both C and C++.

compiler/options.m:
	Add a new option `--generate-standalone-interface' that causes the
	compiler to generate a header/object pair that can be used to
	intialise/shut down the Mercury runtime from a foreign application.
	The basename of the header/object pair is given as an argument to this
	option.

compiler/mercury_compile.m:
	Create the standalone interface if invoked with
	`--generate-standalone-interface'.

	Emit an error message if `--generate-standalone-interface' is
	specified with `--target java' or `--target il'.  We don't
	currently support that.

compiler/compile_target_code.m:
	Add code to implement the `--generate-standalone-interface' option.

	Fix a typo: s/Serarator/Separator/.

	Fix an overlong line.

compiler/handle_options.m:
	Emit an error message if `--generate-standalone-interface' and
	`--extra-inits' are specified together.

compiler/passes_aux.m:
	Fix some broken indentation.

util/mkinit.c:
	Add a new mode of operation that generates standalone interfaces.
	The principle differences between a standalone interface and a
	_init.c file are that the former sets the program entry point to
	MR_dummy_main and does not create a main function.

	The new `-s' option tells mkinit to create a standalone interface.
	(Note that `-s' implies `-l'.)

	Add a comment pointing to various places that may need to updated if
	mkinit.c is changed.

runtime/mercury_wrapper.{h,c}:
	Add a new procedure MR_dummy_main for use with standalone
	interfaces.  Any attempt to call main/2 through the usual entry
	point when operating in standalone mode will cause a runtime abort.