Emit an error message if `mmc --make' is invoked with a target that contains a
Estimated hours taken: 1
Branches: main

Emit an error message if `mmc --make' is invoked with a target that contains a
directory separator. Targets with a directory component (be it a full pathname
or '../') are not handled correctly by the rest of the code. Previously this
lead to an assertion failure in the compiler.

compiler/make.m:
    Check for targets containing directory separator and emit an error message.