Remove an over-zealous sanity check in the standard library initialisation
Estimated hours taken: 0.5
Branches: main

Remove an over-zealous sanity check in the standard library initialisation
code that causes Mercury programs that run in environments in which stdin or
stdout are not available to abort.  It also fixes a problem with running
Mercury programs under recent versions of GNU nohup.

library/io.m:
	Don't abort if we cannot create binary stdin/stdout.

	Fix some formatting.