Add support for thread local trailing in grades that support parallel
Estimated hours taken: 14
Branches: main

Add support for thread local trailing in grades that support parallel
execution.  Previously the trail state for each context was copied into the
global variables defined in mercury_trail.[ch] when the context was loaded,
i.e. we had thread-local trailing in grades that did not support parallel
execution.

This change extends the runtime to support thread local trailing in the
presence of parallel execution.  We do this by adding the relevant fields to
the MercuryEngine structure and redefining the abstract registers related to
trailing to make accesses to the trail state go via the relevant engine.

This also works for parallel trailing grades with the high-level C backend
because in those grades the trail state is stored in a thread-local dummy
engine and context.  (The context, and its attached trail state, are created
when the thread is spawned.)

XXX the coupling between the high-level and low-level versions of the runtimes
w.r.t trailing is not good.  I intend to fix it as a separate change.
(It also affects thread local mutables.)

runtime/mercury_context.h:
	Fix a comment: a context's trail_zone field should be accessed
	via MR_eng_context, not via the an abstract machine register.

runtime/mercury_engine.h:
	Extend engines so that in parallel trailing grades they contain the
	information necessary to keep track of the state of the trail.

runtime/mercury_regs.h:
	In parallel grades MR_trail_ptr, MR_ticket_counter and
	MR_ticket_high_water need to be accessed w.r.t the engine that the
	context is being executed on.

	Modify the definition of MR_restore_trail_registers() so that it
	accesses the trail state via abstract registers rather than
	manipulating the global variables defined in mercury_trail.[ch]
	directly.

runtime/mercury_trail.[ch]:
	Only define global variables to hold the trail state in non-parallel
	grades.

	Document the difference between parallel and non-parallel trailing
	grades.

library/benchmarking.m:
	In .par.tr grades, ML_report_stats should print the size of the trail
	for the current context.

NEWS:
	Announce support for thread local trailing in parallel grades.