Add an optimisation to the Erlang backend.
Branches: main

Add an optimisation to the Erlang backend.  This change delays the calls to
RTTI functions that return type_ctor_infos.  In place of the type_ctor_info
data we pass the thunks that will return type_ctor_infos when called and call
them as required.

In many cases the type_ctor_info won't be needed so the functions won't be
called.  On the other hand, it could be the case that a function which would
only have been called once previously will now be called multiple times.
The Mercury compiler built with the Erlang backend, compiled with HiPE at the
default optimisation level, is ~12x faster after this change.

compiler/elds_to_erlang.m:
	When generating references to type_ctor_infos, output a function
	reference instead of a function call.

library/erlang_rtti_implementation.m:
	Add a new type type_ctor_info_evaled (only used in the Erlang backend)
	to distinguish between actual type_ctor_infos and the thunks.  Change
	some type signatures to use type_ctor_info_evaled.

	Call the thunk when extracting type_ctor_infos from type_infos.

	In create_type_info, wrap the evaluted type_ctor_info data back into a
	thunk when creating a type_info.

library/type_desc.m:
	Make type_ctor_name_and_arity evaluate type_ctor_info thunks.

	Make the Erlang functions that unify and compare type_desc and
	type_ctor_descs evaluate thunks if necessary.

	Make the Erlang unify and compare functions for pseudo_type_descs into
	stubs that throw and exception when called.  (I don't think it works
	and should be written properly when it's needed.)