Change the Erlang code that we generate for an if-then-else with a model_non
Estimated hours taken: 0.5
Branches: main

Change the Erlang code that we generate for an if-then-else with a model_non
condition.  If the condition succeeds, we record that in the process
dictionary (like setting a temporary variable) and erase it from the process
dictionary at the end of the if-then-else.  However, we didn't do the clean
up if the if-then-else was exitted by an exception.  Now we wrap the code of
the if-then-else in a `try' statement and always do the cleanup.

compiler/elds.m:
	Change the ELDS to allow try expressions with an optional `catch'
	pattern and an optional `after' expression (for cleaning up).

compiler/erl_code_gen.m:
	Change the generated code as described above

compiler/elds_to_erlang.m:
compiler/erl_code_util.m:
	Conform to the change in the ELDS.