Fix the failure of hard_coded/ee_dummy in low-level C grades.
Estimated hours taken: 0.5
Branches: main

Fix the failure of hard_coded/ee_dummy in low-level C grades.
It was failing because it tested the value of an exported
enumeration constant for a dummy type against the value of a variable
that corresponds to a foreign_proc input argument of that type.

The low-level C backend was not zeroing out the values of dummy input
arguments to foreign_procs, so they just contained rubbish and the
comparison failed.  The fix is to initialise the values of the variables
that correspond to the foreign_proc dummy input argument to zero.
(The high-level C backend does already does this.

compiler/llds_out.m:
	Initialise the variables that correspond to foreign_proc
	dummy input arguments to zero.