Add an interface to the Boehm collectors finalisation mechanism to the
Estimated hours taken: 0.5
Branches: main

Add an interface to the Boehm collectors finalisation mechanism to the
Mercury runtime.  Mercury bindings to C libraries sometimes
require this facility in order to reclaim memory when some of the
memory is under the control of an external library.
Since we may support alternative garbage collectors in the future, it
is preferable that the Mercury runtime provide this facility rather
than user code directly calling the code in Boehm.

runtime/mercury_memory.h:
	Add a wrapper macro for GC_REGISTER_FINALIZER.